+++ src/components/front/base/FontHeader.vue
@@ -124,7 +124,7 @@
 import { RouterLink } from 'vue-router'
 
 const menuData = [
-  { title: '首頁', to: { name: 'FontIndex' } },
+  { title: '首頁', to: { name: 'FrontIndex' } },
   { title: '步道分級介紹', to: { name: 'TrailsIntro' } },
   { title: '探索步道', to: { name: 'TrailsList' } },
   { title: '郊友護照', to: { name: 'PassportIndex' } }
