+++ src/views/front/FrontIndexView.vue
